requirements.txt
aiogram==2.11.2
aiohttp==3.7.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
chardet==3.0.4
cycler==0.10.0
idna==3.1
kiwisolver==1.3.1
matplotlib==3.3.3
multidict==5.1.0
numpy==1.19.5
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
torch
torchaudio
torchvision
typing-extensions==3.7.4.3
yarl==1.6.3
